Game

+++ app/Controllers/Admin/Game.php
@@ -3,12 +3,15 @@
 namespace App\Controllers\Admin;
 
 use App\Controllers\BaseController;
+use App\Models\ClassificationModel;
 use App\Models\ClubModel;
 use App\Models\DivisionModel;
 use App\Models\GameModel;
 use App\Models\GymModel;
 use App\Models\ServiceModel;
 use App\Models\ServiceGameModel;
+use App\Models\TechnicalFoulModel;
+use App\Models\TypeModel;
 use CodeIgniter\HTTP\ResponseInterface;
 
 class Game extends AdminController
@@ -20,6 +23,10 @@ class Game extends AdminController
 
     protected $serviceModel;
     protected $serviceGameModel;
+    protected $typeModel;
+    protected $classificationModel;
+
+    protected $technicalFoulModel;
 
     public function __construct(){
         $this->clubModel = new ClubModel();
@@ -28,6 +35,9 @@ class Game extends AdminController
         $this->gymModel = new GymModel();
         $this->serviceModel = new ServiceModel();
         $this->serviceGameModel = new ServiceGameModel();
+        $this->typeModel = new TypeModel();
+        $this->classificationModel = new ClassificationModel();
+        $this->technicalFoulModel = new TechnicalFoulModel();
     }
 
     public function index()
@@ -43,11 +53,14 @@ class Game extends AdminController
     public function form($id = null) {
         $this->addBreadcrumb('Liste des matchs', '/admin/game');
         $services = $this->serviceModel->findAll();
+        $typesTF = $this->typeModel->findAll();
+        $classificationsTF = $this->classificationModel->findAll();
 
         if ($id != null) {
             $title = 'Modifier un match';
             $game = $this->gameModel->getFullGame($id);
             $game->services = $this->serviceGameModel->getServicesByGame($id);
+            $game->technical_fouls = $this->technicalFoulModel->where('id_game',$id)->getTechnicalFoulsWithInfos();
         } else {
             $title = 'Ajouter un match';
         }
@@ -56,6 +69,8 @@ class Game extends AdminController
             'title' => $title,
             'game' => $game ?? null,
             'services' => $services ?? null,
+            'typesTF' => $typesTF ?? null,
+            'classificationsTF' => $classificationsTF ?? null,
             ];
         return $this->render('admin/game/form', $data);
     }
@@ -82,6 +97,11 @@ class Game extends AdminController
             $services = $this->request->getPost('services');
             $deletedServices = $this->request->getPost('deletedServices');
 
+            //Récupération des fautes techniques
+            $technicalFouls = $this->request->getPost('technical_fouls');
+            $deletedTechnicalFouls = $this->request->getPost('deletedTechnicalFouls');
+
+
             //Variable pour savoir si c'est un nouveau match
             $newGame=empty($dataGame['id']);
 
@@ -94,31 +114,66 @@ class Game extends AdminController
                 $id=$this->gameModel->getInsertID();
             }
 
-            //GESTION DES SERVICES
+            //START : GESTION DES FAUTES TECHNIQUES
+            //Gestion des suppressions des fautes techniques
+            if(!empty($deletedTechnicalFouls)){
+                $existingTechnicalFoulsId = $this->technicalFoulModel->where('id_game',$id)->findcolumn('id');
+                foreach ($deletedTechnicalFouls as $deletedTF){
+                    $idTF = $deletedTF ['id'];
+                    if(in_array($idTF,$existingTechnicalFoulsId)){
+                        $this->technicalFoulModel->delete($idTF);
+                    }
+                }
+            }
+
+            //Enregistrement des fautes techniques en BDD
+            if(!empty($technicalFouls)){
+                foreach ($technicalFouls as $technicalFoul) {
+                    $dataTechnicalFoul = [
+                        'id' => $technicalFoul['id'] ?? null,
+                        'id_game' => $id,
+                        'id_member' => $technicalFoul ['player'],
+                        'id_type' => $technicalFoul ['type'],
+                        'id_classification' => $technicalFoul ['classification'],
+                        'amount' => $technicalFoul ['amount'],
+                    ];
+                    if(!$this->technicalFoulModel->save($dataTechnicalFoul)){
+                        $this->error(implode('<br>',$this->technicalFoulModel->errors()));
+                    }
+                }
+            }
+            //END : GESTION DES FAUTES TECHNIQUES
+
+            //START : GESTION DES SERVICES
             //Récupération des services existants pour ce match
             $existingServices = $this->serviceGameModel->where('id_game', $id)->findAll();
 
-            //On crée un équivalent de la clé composite (sans_id_game car déjà filtré)
-            $existingKeys = [];
+            //On crée les équivalents de clés composites (sans_id_game car déjà filtré)
+            $serviceKeys = [];
+            if($services){
+                foreach($services as $service){
+                    $serviceKeys[] = $service['id_service'].'-'.$service['id_member'];
+                }
+            }
+
+            $existingServiceKeys = [];
             if($existingServices){
                 foreach($existingServices as $existingService){
-                    $existingKeys[] = $existingService['id_service'].'-'.$existingService['id_member'];
+                    $key = $existingService['id_service'].'-'.$existingService['id_member'];
+                    if(!in_array($key, $serviceKeys)){
+                        if(!$this->serviceGameModel->where('id_service', $existingService['id_service'])->where('id_game',$id)->where('id_member',$existingService['id_member'])->delete()){
+                            $this->error(implode('<br>',$this->serviceGameModel->errors()));
+                        }
+                    }
+                    $existingServiceKeys[] = $key;
                 }
             }
 
-
-            //suppression des services supprimés
-            $deletedKeys = [];
-            if($deletedServices){
+            //Suppression des services supprimés
+            if($deletedServices && !empty($deletedServices)){
                 foreach($deletedServices as $deletedService){
-                    $deletedKeys[] = $deletedService['id_service'].'-'.$deletedService['id_member'];
-                }
-            }
-
-            if(!empty($deletedServices)){
-                foreach($deletedServices as $deletedService){
-                    $key = $deletedService['id_service'].'-'.$deletedService['id_member'];
-                    if(in_array($key, $existingKeys)){
+                    $deletedKey = $deletedService['id_service'].'-'.$deletedService['id_member'];
+                    if(in_array($deletedKey, $existingServiceKeys)){
                         if(!$this->serviceGameModel->where('id_service', $deletedService['id_service'])->where('id_game',$id)->where('id_member',$deletedService['id_member'])->delete()){
                             $this->error(implode('<br>',$this->serviceGameModel->errors()));
                         }
@@ -130,7 +185,8 @@ class Game extends AdminController
             if(!empty($services)){
                 foreach($services as $service){
                     $key=$service['id_service'].'-'.$service['id_member'];
-                    if(!in_array($key, $existingKeys)){
+
+                    if(!in_array($key, $existingServiceKeys)){
                         $dataService = [
                             'id_service' => $service['id_service'],
                             'id_game' => $id,
@@ -143,7 +199,7 @@ class Game extends AdminController
                     }
                 }
             }
-
+            //END : GESTION DES SERVICES
 
             //Messages de validation et récupération nouvel ID
             if($newGame){
